Cast to int for printf

+++ main.cpp
@@ -364,9 +364,9 @@ int main()
 
                     break;
                 case MessageType::Text:
-                    char* buf = new char[m.size];
+                    char* buf = new char[(int)m.size];
                     int ret = read(sock, &buf, m.size);
-                    printf("Text event, expected buf size: %zu, received: %d", m.size, ret);
+                    printf("Text event, expected buf size: %d, received: %d", (int)m.size, ret);
                     printf("-> Server: \"%s\"\n", buf);
                     break;
             }
